+++ backend/src/main/java/com/bungeobbang/backend/opinion/presentation/AdminOpinionController.java
@@ -4,7 +4,7 @@ import com.bungeobbang.backend.auth.admin.AdminAuth;
 import com.bungeobbang.backend.auth.admin.AdminOnly;
 import com.bungeobbang.backend.auth.domain.Accessor;
 import com.bungeobbang.backend.opinion.dto.response.AdminOpinionInfoListResponse;
-import com.bungeobbang.backend.opinion.service.OpinionService;
+import com.bungeobbang.backend.opinion.service.AdminOpinionService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -17,13 +17,13 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/admin/opinion")
 public class AdminOpinionController {
 
-    private final OpinionService opinionService;
+    private final AdminOpinionService adminOpinionService;
 
     @GetMapping()
     @AdminOnly
     public ResponseEntity<AdminOpinionInfoListResponse> getAdminOpinionList(
             final @AdminAuth Accessor accessor) {
         return ResponseEntity.ok()
-                .body(opinionService.findAdminOpinionList());
+                .body(adminOpinionService.findAdminOpinionList());
     }
 }
