+++ backend/src/main/java/com/bungeobbang/backend/opinion/service/MemberOpinionService.java
@@ -28,7 +28,7 @@ import java.util.List;
 @Service
 @RequiredArgsConstructor
 @Slf4j
-public class OpinionService {
+public class MemberOpinionService {
 
     private final OpinionRepository opinionRepository;
     private final OpinionChatRepository opinionChatRepository;
@@ -107,44 +107,6 @@ public class OpinionService {
         final List<Opinion> opinions = opinionRepository.findAllByMemberId(memberId);
         final List<MemberOpinionInfoResponse> opinionInfos = convertToMemberOpinionInfoList(opinions);
         return new MemberOpinionInfoListResponse(opinionInfos);
-    }
-
-    /**
-     * 말해요의 전체 채팅방 리스트를 조회합니다.
-     *
-     * @return AdminOpinionInfoListResponse 모든 말해요 채팅방 정보 리스트
-     */
-    public AdminOpinionInfoListResponse findAdminOpinionList() {
-        final List<Opinion> opinions = opinionRepository.findAll();
-        final List<AdminOpinionInfoResponse> opinionInfos = convertToAdminOpinionInfoList(opinions);
-        return new AdminOpinionInfoListResponse(opinionInfos);
-    }
-
-    private List<AdminOpinionInfoResponse> convertToAdminOpinionInfoList(final List<Opinion> opinions) {
-        return opinions.stream()
-                .map(opinion -> {
-                    log.debug("opinionId: {}", opinion.getId());
-
-                    final OpinionLastRead opinionLastRead = opinionLastReadRepository.findByOpinionIdAndIsAdmin(opinion.getId(), false)
-                            .orElseThrow(() -> new OpinionException(ErrorCode.INVALID_OPINION_LAST_READ));
-                    log.debug("opinionLastReadChatId: {}", opinionLastRead.getLastReadChatId());
-
-                    final OpinionChat lastChat = opinionChatRepository.findTopByOpinionIdOrderByCreatedAtDesc(opinion.getId())
-                            .orElseThrow(() -> new OpinionException(ErrorCode.INVALID_OPINION_CHAT));
-                    log.debug("lastChatId: " + lastChat.getId());
-
-                    return AdminOpinionInfoResponse.builder()
-                            .opinionId(opinion.getId())
-                            .opinionType(opinion.getOpinionType())
-                            .categoryType(opinion.getCategoryType())
-                            .lastChat(lastChat.getChat())
-                            .lastChatCreatedAt(lastChat.getCreatedAt())
-                            .isNew(!opinionLastRead.getLastReadChatId().equals(lastChat.getId()))
-                            .isRemind(opinion.isRemind())
-                            .build();
-                })
-                .sorted()
-                .toList();
     }
 
     /**
